udgoap working
I have no strength left in my body, my ear is numb to the sound of the world, but it is done.
udgoap npc behaves like a npc should
also some adjustments on the goap that was using smart objects and cannot.
also some adjustments to prepare for the testing and collecting of results